+++ react/components/SearchBar/components/SearchBar.js
@@ -110,9 +110,8 @@ const SearchBar = ({
         const paramForSearchTracking = '&_c=' + terms[0]
 
         page = 'store.search'
-        params = { term: terms[0] }
-        query =
-          `map=c,ft&rest=${terms.slice(1).join(',')}` + paramForSearchTracking
+        params = { term: terms.join('/') }
+        query = `map=c,ft` + paramForSearchTracking
       }
 
       navigate({ page, params, query })